+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 - Added 52 new colours and 66 new animals, taking the total number of combinations
   from ~33 million to ~64 million. This also increases the maximum length of an ACA
   from 40 characters to 41 (assuming a single character delimiter).
+- Added `NewDefaultGenerator()` factory that uses a default delimiter (`-`) and
+  source of randomness.